+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 - `react/forbid-component-props` new rule from eslint-plugin-react v6.1.0
 - `react/jsx-filename-extension` new rule from eslint-plugin-react v5.2.0
 - `react/jsx-no-target-blank` new rule from eslint-plugin-react v5.1.0
+- `react/jsx-tag-spacing` new rule from eslint-plugin-react v6.7.0
 - `react/no-children-prop` new rule from eslint-plugin-react v6.3.0
 - `react/no-comment-textnodes` new rule from eslint-plugin-react v5.2.0
 - `react/no-danger-with-children` new rule from eslint-plugin-react v6.1.0